Code refactor - split big modules into smaller ones
