+++ newsfragments/452.misc.rst
@@ -1,0 +1,1 @@
+Split bigger code modules into smaller chunks.
