+++ pytest_redis/executor/__init__.py
@@ -1,0 +1,6 @@
+"""Redis executor."""
+
+from pytest_redis.executor.process import RedisExecutor
+from pytest_redis.executor.noop import NoopRedis
+
+__all__ = ("RedisExecutor", "NoopRedis")
