+++ pytest_redis/factories/__init__.py
@@ -1,0 +1,4 @@
+"""Redis fixture factories."""
+from pytest_redis.factories.proc import redis_proc
+from pytest_redis.factories.noproc import redis_noproc
+from pytest_redis.factories.client import redisdb
