+++ tests/conftest.py
@@ -1,6 +1,9 @@
 """Tests main conftest file."""
 import warnings
 
+import pytest_redis.factories.client
+import pytest_redis.factories.noproc
+import pytest_redis.factories.proc
 from pytest_redis.plugin import *
 from pytest_redis import factories
 
@@ -9,13 +12,13 @@ warnings.filterwarnings(
 )
 
 # pylint:disable=invalid-name
-redis_proc2 = factories.redis_proc(port=6381)
-redis_nooproc2 = factories.redis_noproc(port=6381, startup_timeout=1)
-redis_proc3 = factories.redis_proc(port=6385, password="secretpassword")
-redis_nooproc3 = factories.redis_noproc(port=6385, password="secretpassword")
+redis_proc2 = pytest_redis.factories.proc.redis_proc(port=6381)
+redis_nooproc2 = pytest_redis.factories.noproc.redis_noproc(port=6381, startup_timeout=1)
+redis_proc3 = pytest_redis.factories.proc.redis_proc(port=6385, password="secretpassword")
+redis_nooproc3 = pytest_redis.factories.noproc.redis_noproc(port=6385, password="secretpassword")
 
-redisdb2 = factories.redisdb("redis_proc2")
-redisdb2_noop = factories.redisdb("redis_nooproc2")
-redisdb3 = factories.redisdb("redis_proc3")
-redisdb3_noop = factories.redisdb("redis_nooproc3")
+redisdb2 = pytest_redis.factories.client.redisdb("redis_proc2")
+redisdb2_noop = pytest_redis.factories.client.redisdb("redis_nooproc2")
+redisdb3 = pytest_redis.factories.client.redisdb("redis_proc3")
+redisdb3_noop = pytest_redis.factories.client.redisdb("redis_nooproc3")
 # pylint:enable=invalid-name
